feat: custom hook for rubric generator
