+++ templates/Rubric/RubricGenerator.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React  from "react";
 import {
   Box,
   Button,
@@ -11,87 +11,30 @@ import {
   InputAdornment,
   IconButton,
 } from "@mui/material";
-import axios from "axios";
-import { useSelector } from "react-redux";
 import { styles } from "./styles";
 import { ArrowDropDown, UploadFile } from "@mui/icons-material";
 import FileUploadDialog from "./FileUploadDialog";
+import useRubricGenerator from '@/libs/hooks/useRubricGenerator';
+
 
 const RubricGenerator = ({ onGenerateRubric }) => {
-  const user = useSelector((state) => state.user.data);
-  const [gradeLevel, setGradeLevel] = useState("");
-  const [pointScale, setPointScale] = useState("");
-  const [standards, setStandards] = useState("");
-  const [assignmentDescription, setAssignmentDescription] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState("");
-  const [openFileUploadDialog, setOpenFileUploadDialog] = useState(false);
-  const [objectivesFileUrl, setObjectivesFileUrl] = useState("");
-  const [assignmentDescriptionFileUrl, setAssignmentDescriptionFileUrl] = useState("");
-
-  const handleFileUploadClick = () => {
-    setOpenFileUploadDialog(true);
-  };
-
-  const handleFileUpload = (fileType, fileUrl) => {
-    if (fileType === "objectives") {
-      setObjectivesFileUrl(fileUrl);
-    } else if (fileType === "assignment_description") {
-      setAssignmentDescriptionFileUrl(fileUrl);
-    }
-  };
-
-
-  const headers = {
-    "API-Key": "dev",
-    "Content-Type": "application/json",
-  };
-
-  const handleGenerateRubric = async () => {
-    if (!gradeLevel || !pointScale || !assignmentDescription) {
-      setError("All fields are required.");
-      return;
-    }
-
-    const payload = {
-      user: user,
-      type: "chat",
-      tool_data: {
-        tool_id: "rubric-generator",
-        inputs: [
-          { name: "grade_level", value: gradeLevel },
-          { name: "point_scale", value: parseInt(pointScale) },
-          { name: "objectives", value: standards },
-          { name: "assignment_description", value: assignmentDescription },
-          { name: "objectives_file_url", value: objectivesFileUrl },
-          { name: "objectives_file_type", value: "pdf" },
-          { name: "assignment_description_file_url", value: assignmentDescriptionFileUrl },
-          { name: "assignment_description_file_type", value: "gdoc" },
-          { name: "lang", value: "en" },
-        ],
-      },
-    };
-
-    try {
-      setError("");
-      setLoading(true);
-      const response = await axios.post(
-        `${process.env.NEXT_PUBLIC_MARVEL_ENDPOINT}submit-tool`,
-        payload,
-        { headers }
-      );
-      onGenerateRubric(response.data);
-    } catch (err) {
-      console.error("Error response:", err.response || err.message);
-      if (err.response?.data?.message) {
-        setError(err.response.data.message);
-      } else {
-        setError("Failed to generate rubric. Please try again later.");
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
+  const {
+    gradeLevel,
+    setGradeLevel,
+    pointScale,
+    setPointScale,
+    standards,
+    setStandards,
+    assignmentDescription,
+    setAssignmentDescription,
+    loading,
+    error,
+    openFileUploadDialog,
+    setOpenFileUploadDialog,
+    handleFileUploadClick,
+    handleFileUpload,
+    handleGenerateRubric,
+  } = useRubricGenerator();
 
   return (
     <>
@@ -211,7 +154,7 @@ const RubricGenerator = ({ onGenerateRubric }) => {
         >
           <Button
             disabled={loading}
-            onClick={handleGenerateRubric}
+            onClick={() => handleGenerateRubric(onGenerateRubric)}
             sx={styles.generateButton}
           >
             Generate
